Move parts of typecheck into new typecheck::unif
The typecheck module is several kLoC. This PR creates a new submodule
`typecheck::unif`, which focus on unification internals, as to make those modules
more digestible.
